interpreter: move RawPtr and TypedPtr to own file
